requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# WebSocket Support
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0

# Database
psycopg2-binary==2.9.9

# Security & Encryption
cryptography==41.0.7
pynacl==1.5.0
argon2-cffi==23.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
redis==5.0.1
gunicorn==21.2.0

# For secure key derivation
scrypt==0.8.21
